Set time ok if button is pressed for two seconds
Since a timekeeper with a backup battery is used now to count the time,
there is no way to check whether the time is correct after the main
power was switched on. (You cannot simply check if the time was reset
to its initial value since the time just freezes if the backup battery
was removed or run out of energy.) So the clocks colon starts to blick
after powering up and the user need to tell that the time is fine by
pressing and holding the button for two seconds. If the user presses
and holds the button for two seconds again, the colon starts to blink
again. This is usefull the check if an DCF77 telegram was received
successfully since the colon, as before, stops to blink in this case.